Split up compose ts
