+++ src/ifcx-core/composition/path.ts
@@ -1,0 +1,12 @@
+
+export function GetHead(path: string)
+{
+    return path.split("/")[0];
+}
+
+export function GetTail(path: string)
+{
+    let parts = path.split("/");
+    parts.shift();
+    return parts.join("/");
+}
