+++ src/ifcx-core/schema/schema-helper.ts
@@ -1,7 +1,7 @@
 // the current typespec -> openapi -> TS output has a strange structure
 // TODO: use typespec TS export and remove this file
 
-import { components } from "../../schema/out/ts/ifcx";
+import { components } from "../../../schema/out/ts/ifcx";
 
 export type IfcxFile = components["schemas"]["IfcxFile"];
 export type IfcxSchema = components["schemas"]["IfcxSchema"];
