+++ src/ifcx-core/util/mm.ts
@@ -1,0 +1,13 @@
+
+// this is a helper function that makes a regular Map behave as a multi map
+export function MMSet<A, B>(map: Map<A, B[]>, key: A, value: B)
+{
+    if (map.has(key))
+    {
+        map.get(key)?.push(value);
+    }
+    else
+    {
+        map.set(key, [value]);
+    }
+}
