+++ src/test/schema-test.ts
@@ -1,5 +1,5 @@
-import { SchemasToOpenAPI } from "../ifcx-core/schema-export";
-import { SchemaValidationError } from "../ifcx-core/schema-validation";
+import { SchemasToOpenAPI } from "../ifcx-core/schema/schema-export";
+import { SchemaValidationError } from "../ifcx-core/schema/schema-validation";
 import { LoadIfcxFile } from "../ifcx-core/workflows";
 import { ExampleFile, ExampleFileMissingSchema, ExampleFileWithSchema } from "./example-file";
 import { describe, it } from "./util/cappucino";
